raspberry-pi/rpi-seq/requirements.txt
scramjet-framework-py==0.10
pyee==9.0.4
NewlineJSON